new FRACTAL_API_V1_MODE with test
