+++ fractal_server/main.py
@@ -42,12 +42,12 @@ def collect_routers(app: FastAPI) -> None:
     settings = Inject(get_settings)
 
     app.include_router(router_api, prefix="/api")
-    if settings.FRACTAL_API_V1_ENABLED is True:
+    if settings.FRACTAL_API_V1_MODE == "include":
         app.include_router(router_api_v1, prefix="/api/v1")
+        app.include_router(
+            router_admin_v1, prefix="/admin/v1", tags=["V1 Admin area"]
+        )
     app.include_router(router_api_v2, prefix="/api/v2")
-    app.include_router(
-        router_admin_v1, prefix="/admin/v1", tags=["V1 Admin area"]
-    )
     app.include_router(
         router_admin_v2, prefix="/admin/v2", tags=["V2 Admin area"]
     )
