+++ tests/conftest.py
@@ -66,3 +66,4 @@ from .fixtures_server_v2 import *  # noqa F403
 from .fixtures_tasks_v1 import *  # noqa F403
 from .fixtures_slurm import *  # noqa F403
 from .fixtures_tasks_v2 import *  # noqa F403
+from .fixtures_commands import *  # noqa F403
